Floats to doubles everywhere. Removal of NoveltyFitness. Refractoring of post-evals and scoring mechanisms to be more general. Some classes moved around. Only basic tests were done.

+++ src/mase/app/aggregation/AggregationBehaviourEval.java
@@ -15,12 +15,12 @@ import sim.util.MutableDouble2D;
  */
 public class AggregationBehaviourEval extends MasonEvaluation {
 
-    private float[] centerMass;
+    private double[] centerMass;
     private VectorBehaviourResult vbr;
 
     @Override
     protected void preSimulation() {
-        this.centerMass = new float[super.maxEvaluationSteps];
+        this.centerMass = new double[super.maxEvaluationSteps];
     }
 
     @Override
@@ -36,7 +36,7 @@ public class AggregationBehaviourEval extends MasonEvaluation {
             dist += centre.distance(aa.getLocation());
         }
         dist = dist / agg.agents.size() / agg.par.size;
-        centerMass[super.currentEvaluationStep] = (float) dist;
+        centerMass[super.currentEvaluationStep] = dist;
     }
 
     @Override
